set up fastapi
