+++ requirements.txt
@@ -2,6 +2,7 @@ aiohappyeyeballs==2.6.1
 aiohttp==3.12.7
 aiosignal==1.3.2
 annotated-types==0.7.0
+anyio==4.9.0
 async-timeout==5.0.1
 attrs==25.3.0
 blessed==1.21.0
@@ -9,12 +10,25 @@ certifi==2025.4.26
 cffi==1.17.1
 chacha20poly1305-reuseable==0.13.2
 charset-normalizer==3.4.2
+click==8.2.1
 cryptography==45.0.3
+dnspython==2.7.0
 editor==1.6.6
+email_validator==2.2.0
+fastapi==0.115.12
+fastapi-cli==0.0.7
 frozenlist==1.6.0
+h11==0.16.0
+httpcore==1.0.9
+httptools==0.6.4
+httpx==0.28.1
 idna==3.10
 ifaddr==0.2.0
 inquirer==3.4.0
+Jinja2==3.1.6
+markdown-it-py==3.0.0
+MarkupSafe==3.0.2
+mdurl==0.1.2
 miniaudio==1.61
 multidict==6.4.4
 propcache==0.3.1
@@ -23,17 +37,31 @@ pyatv==0.16.0
 pycparser==2.22
 pydantic==2.11.5
 pydantic_core==2.33.2
+Pygments==2.19.1
+python-dotenv==1.1.0
+python-multipart==0.0.20
+PyYAML==6.0.2
 readchar==4.2.1
 requests==2.32.3
+rich==14.0.0
+rich-toolkit==0.14.7
 runs==1.2.2
+shellingham==1.5.4
 six==1.17.0
+sniffio==1.3.1
 srptools==1.0.1
+starlette==0.46.2
 tabulate==0.9.0
 tinytag==2.1.1
+typer==0.16.0
 typing-inspection==0.4.1
 typing_extensions==4.14.0
 urllib3==2.4.0
+uvicorn==0.34.3
+uvloop==0.21.0
+watchfiles==1.0.5
 wcwidth==0.2.13
+websockets==15.0.1
 xmod==1.8.1
 yarl==1.20.0
 zeroconf==0.147.0
